refactor: add monitor loop package
Add loop package for monitors.
Gofmt/goimports source code.
